day_10/server/requirements.txt
fastmcp==2.13.0.2
python-dotenv==1.1.1
httpx>=0.27
pydantic>=2.7
jinja2>=3.1
python-slugify>=8.0
